jose::algorithm

Enum Algorithm

pub enum Algorithm {

    HS256,
    HS384,
    HS512,
    RS256,
    RS384,
    RS512,
    ES256,
    ES384,
    ES512,
    ES256K,
    PS256,
    PS384,
    PS512,
    EdDSA,
    Ed25519,
    Ed448,
    MLDSA44,
    MLDSA65,
    MLDSA87,
    None,
    RSA1_5,
    RSAOAEP,
    RSAOAEP256,
    RSAOAEP384,
    RSAOAEP512,
    A128KW,
    A192KW,
    A256KW,
    Dir,
    ECDHES,
    ECDHESA128KW,
    ECDHESA192KW,
    ECDHESA256KW,
    A128GCMKW,
    A192GCMKW,
    A256GCMKW,
    PBES2HS256A128KW,
    PBES2HS384A192KW,
    PBES2HS512A256KW,

}

JSON Web Signature and Encryption Algorithms

Based on the IANA JSON Web Signature and Encryption Algorithms registry. See: https://www.iana.org/assignments/jose/jose.xhtml#web-signature-encryption-algorithms

Variants

HS256

HS384

HS512

RS256

RS384

RS512

ES256

ES384

ES512

ES256K

PS256

PS384

PS512

EdDSA

Ed25519

Ed448

MLDSA44

MLDSA65

MLDSA87

None

RSA1_5

RSAOAEP

RSAOAEP256

RSAOAEP384

RSAOAEP512

A128KW

A192KW

A256KW

Dir

ECDHES

ECDHESA128KW

ECDHESA192KW

ECDHESA256KW

A128GCMKW

A192GCMKW

A256GCMKW

PBES2HS256A128KW

PBES2HS384A192KW

PBES2HS512A256KW

Implementations

impl Algorithm

pub fn is_signature_algorithm(&self) -> bool

Returns true if this algorithm is suitable for JWS (signing)

pub fn is_key_encryption_algorithm(&self) -> bool

Returns true if this algorithm is suitable for JWE key encryption

pub fn requires_additional_params(&self) -> bool

Returns true if this algorithm requires additional parameters

pub fn recommended_key_size(&self) -> Option<usize>

Returns the recommended key size in bits for this algorithm